requirements.txt
pywin32
beautifulsoup4
selenium
python-dotenv
pytz